Changes to eliminate compiler warnings.  The biggest change is to the struct uint4.  It was a struct of 4 unsigned ints.  It is now a struct of one unsigned int and a struct taus_state_t, which is a struct of three unsigned ints.  This change means that the Tausworthe functions don't need to receive void pointers, which was causing strict-aliasing warnings.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@6426 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/io/PVParams.cpp
@@ -1279,7 +1279,7 @@ const char * PVParams::getFilename(const char * id)
  */
 void PVParams::addGroup(char * keyword, char * name)
 {
-   assert(numGroups <= groupArraySize);
+   assert((size_t) numGroups <= groupArraySize);
 
    // Verify that the new group's name is not an existing group's name
    for( int k=0; k<numGroups; k++ ) {
